// packages/uimodel-composer/src/generated/Expression.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

import type { EObject, BasicEObject, EClass, EStructuralFeature } from '@emfts/core';

/**
 * Expression
 * @generated
 */
export interface Expression extends EObject {
  language: string;
  body: string;

}